+++ src/test/java/stepdefinitions/GraphModule.java
@@ -1,0 +1,54 @@
+package stepdefinitions;
+
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class GraphModule {
+	@When("The user write code in Editor from sheetname {string} and rownumber {int}")
+	public void the_user_write_code_in_editor_from_sheetname_and_rownumber(String string, Integer int1) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@When("Click the run button")
+	public void click_the_run_button() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user should able to see output in the console {string}")
+	public void the_user_should_able_to_see_output_in_the_console(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user should able to see an error message in alert window")
+	public void the_user_should_able_to_see_an_error_message_in_alert_window() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@When("The user clicks on ok button in alert window")
+	public void the_user_clicks_on_ok_button_in_alert_window() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user is on the {string}  Page")
+	public void the_user_is_on_the_page(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@When("User click back button")
+	public void user_click_back_button() {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+	}
+
+	@Then("The user should navigate to URL {string}")
+	public void the_user_should_navigate_to_url(String string) {
+	    // Write code here that turns the phrase above into concrete actions
+	    throw new io.cucumber.java.PendingException();
+}
+}
