+++ src/test/java/stepdefinitions/Tree.java
@@ -18,16 +18,16 @@ public class Tree {
 	}
 
 	@Given("The user is on the {string} Home Page")
-	public void the_user_is_on_the_home_page1(String string) {
+	//public void the_user_is_on_the_home_page1(String string) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	   // throw new io.cucumber.java.PendingException();
+	//}
 
 	@When("User clicks on Get Started button index {int}")
-	public void user_clicks_on_get_started_button_index(Integer int1) {
+	//public void user_clicks_on_get_started_button_index(Integer int1) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	  //  throw new io.cucumber.java.PendingException();
+	//}
 
 	@Then("User should be on {string} page")
 	public void user_should_be_on_page(String string) {
@@ -60,50 +60,50 @@ public class Tree {
 	}
 
 	@When("The user is on the {string} Home Page")
-	public void the_user_is_on_the_home_page(String string) {
+	//public void the_user_is_on_the_home_page(String string) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	//    throw new io.cucumber.java.PendingException();
+	//}
 
 	@When("The user write code in Editor from sheetname {string} and rownumber {int}")
-	public void the_user_write_code_in_editor_from_sheetname_and_rownumber(String string, Integer int1) {
+	//public void the_user_write_code_in_editor_from_sheetname_and_rownumber(String string, Integer int1) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	 //   throw new io.cucumber.java.PendingException();
+	//}
 
 	@When("Click the run button")
-	public void click_the_run_button() {
+	//public void click_the_run_button() {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	   // throw new io.cucumber.java.PendingException();
+	//}
 
 	@Then("The user should able to see output in the console {string}")
-	public void the_user_should_able_to_see_output_in_the_console(String string) {
+	//public void the_user_should_able_to_see_output_in_the_console(String string) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	  //  throw new io.cucumber.java.PendingException();
+	//}
 
 	@Then("The user is on the {string} Home Page")
-	public void the_user_is_on_the_home_page1(String string) {
+	//public void the_user_is_on_the_home_page1(String string) {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	 //   throw new io.cucumber.java.PendingException();
+	//}
 
 	@Then("The user should able to see an error message in alert window")
-	public void the_user_should_able_to_see_an_error_message_in_alert_window() {
+	//public void the_user_should_able_to_see_an_error_message_in_alert_window() {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	  //  throw new io.cucumber.java.PendingException();
+	//}
 
 	@When("The user clicks on Ok button in alert window")
-	public void the_user_clicks_on_ok_button_in_alert_window() {
+	//public void the_user_clicks_on_ok_button_in_alert_window() {
 	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	    //throw new io.cucumber.java.PendingException();
+	//}
 
 	@Then("The user should be on same page")
-	public void the_user_should_be_on_same_page() {
-	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
+	//public void the_user_should_be_on_same_page() {
+	     //Write code here that turns the phrase above into concrete actions
+	  //  throw new io.cucumber.java.PendingException();
+	//}
 }
