+++ src/test/resources/feature/Signout2.feature
@@ -1,0 +1,46 @@
+#Author: your.email@your.domain.com
+#Keywords Summary :
+#Feature: List of scenarios.
+#Scenario: Business rule through list of steps with arguments.
+#Given: Some precondition step
+#When: Some key actions
+#Then: To observe outcomes or validation
+#And,But: To enumerate more Given,When,Then steps
+#Scenario Outline: List of steps for data-driven as an Examples and <placeholder>
+#Examples: Container for s table
+#Background: List of steps run before each of the scenarios
+#""" (Doc Strings)
+#| (Data Tables)
+#@ (Tags/Labels):To group Scenarios
+#<> (placeholder)
+#""
+## (Comments)
+#Sample Feature Definition Template
+@tag
+Feature: Checking Sign Out
+Background:
+     Given User is  in DSAlgo Portal
+     When User sign In to portal 
+  @tag1
+  Scenario: Check Sign out Button
+  
+    Given User is in DSAlgo Home page after signed in
+    
+    
+    When User clicks on Sign Out
+    
+    
+    Then The User shoud see the Logged out successfully message in the home page  with Register and Sign in Options
+    
+    
+
+  #@tag2
+  #Scenario Outline: Title of your scenario outline
+    #Given I want to write a step with <name>
+    #When I check for the <value> in step
+    #Then I verify the <status> in step
+
+    #Examples: 
+     # | name  | value | status  |
+     # | name1 |     5 | success |
+     # | name2 |     7 | Fail    |
